tests/wpt/web-platform-tests/shadow-dom/untriaged/html-elements-and-their-shadow-trees/test-001.html

<!DOCTYPE html>
<!--
Distributed under both the W3C Test Suite License [1] and the W3C
3-clause BSD License [2]. To contribute to a W3C Test Suite, see the
policies and contribution forms [3].

[1] http://www.w3.org/Consortium/Legal/2008/04-testsuite-license
[2] http://www.w3.org/Consortium/Legal/2008/03-bsd-license
[3] http://www.w3.org/2004/10/27-testcases
-->
<html>
<head>
<title>Shadow DOM Test: A_09_00_01</title>
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="help" href="http://www.w3.org/TR/2013/WD-shadow-dom-20130514/#html-elements-and-their-shadow-trees">
<meta name="assert" content="HTML Elements and Their Shadow Trees: If the element can have fallback content, UA should allow the shadow tree to contain at least one insertion point.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
//test iframe
test(unit(function (ctx) {

	var d = newRenderedHTMLDocument(ctx);

	// create element
    var el = d.createElement('iframe');
    d.body.appendChild(el);

    el.innerHTML = '' +
    	'<span id="shadow">This is a node that should be distributed</span>' +
    	'<span id="flbk">This is a fallback content</span>';

    var s = el.createShadowRoot();
    s.innerHTML = '<content select="#shadow"></content>';

    assert_true(d.querySelector('#shadow').offsetTop > 0, 'Iframe should allow at least one insertion point');
    assert_equals(d.querySelector('#flbk').offsetTop, 0, 'Fallback content shouldn\'t be rendered');

}), 'A_09_00_01_T01');


//test object
test(unit(function (ctx) {

	var d = newRenderedHTMLDocument(ctx);

	// create element
    var el = d.createElement('object');
    d.body.appendChild(el);

    el.innerHTML = '' +
    	'<span id="shadow">This is a node that should be distributed</span>' +
    	'<span id="flbk">This is a fallback content</span>';

    var s = el.createShadowRoot();
    s.innerHTML = '<content select="#shadow"></content>';

    assert_true(d.querySelector('#shadow').offsetTop > 0, 'object should allow at least one insertion point');
    assert_equals(d.querySelector('#flbk').offsetTop, 0, 'Fallback content shouldn\'t be rendered');

}), 'A_09_00_01_T02');


//test video
test(unit(function (ctx) {

	var d = newRenderedHTMLDocument(ctx);

	// create element
    var el = d.createElement('video');
    d.body.appendChild(el);

    el.innerHTML = '' +
    	'<span id="shadow">This is a node that should be distributed</span>' +
    	'<span id="flbk">This is a fallback content</span>';

    var s = el.createShadowRoot();
    s.innerHTML = '<content select="#shadow"></content>';

    assert_true(d.querySelector('#shadow').offsetTop > 0, 'video should allow at least one insertion point');
    assert_equals(d.querySelector('#flbk').offsetTop, 0, 'Fallback content shouldn\'t be rendered');

}), 'A_09_00_01_T03');


//test audio
test(unit(function (ctx) {

	var d = newRenderedHTMLDocument(ctx);

	// create element
    var el = d.createElement('audio');
    d.body.appendChild(el);

    el.innerHTML = '' +
    	'<span id="shadow">This is a node that should be distributed</span>' +
    	'<span id="flbk">This is a fallback content</span>';

    var s = el.createShadowRoot();
    s.innerHTML = '<content select="#shadow"></content>';

    assert_true(d.querySelector('#shadow').offsetTop > 0, 'audio should allow at least one insertion point');
    assert_equals(d.querySelector('#flbk').offsetTop, 0, 'Fallback content shouldn\'t be rendered');

}), 'A_09_00_01_T04');


//test canvas
test(unit(function (ctx) {

	var d = newRenderedHTMLDocument(ctx);

	// create element
    var el = d.createElement('canvas');
    d.body.appendChild(el);

    el.innerHTML = '' +
    	'<span id="shadow">This is a node that should be distributed</span>' +
    	'<span id="flbk">This is a fallback content</span>';

    var s = el.createShadowRoot();
    s.innerHTML = '<content select="#shadow"></content>';

    assert_true(d.querySelector('#shadow').offsetTop > 0, 'canvas should allow at least one insertion point');
    assert_equals(d.querySelector('#flbk').offsetTop, 0, 'Fallback content shouldn\'t be rendered');

}), 'A_09_00_01_T05');


//test map
test(unit(function (ctx) {

	var d = newRenderedHTMLDocument(ctx);

    var img = d.createElement('img');
    img.setAttribute('usemap', '#theMap');
    img.setAttribute('width', '20px');
    img.setAttribute('height', '20px');
    d.body.appendChild(img);


	// create element
    var el = d.createElement('map');
    el.setAttribute('name', 'theMap');
    d.body.appendChild(el);

    el.innerHTML = '' +
		'<span id="shadow">This is a node that should be distributed</span>' +
		'<span id="flbk">This is a fallback content</span>';


    var s = el.createShadowRoot();
    s.innerHTML = '<content select="#shadow"></content>';

    assert_true(d.querySelector('#shadow').offsetTop > 0, 'map should allow at least one insertion point');
    assert_equals(d.querySelector('#flbk').offsetTop, 0, 'Fallback content shouldn\'t be rendered');

}), 'A_09_00_01_T06');


//test textarea
test(unit(function (ctx) {

	var d = newRenderedHTMLDocument(ctx);

	// create element
    var el = d.createElement('textarea');
    d.body.appendChild(el);

    el.innerHTML = '' +
    	'<span id="shadow">This is a node that should be distributed</span>' +
    	'<span id="flbk">This is a fallback content</span>';

    var s = el.createShadowRoot();
    s.innerHTML = '<content select="#shadow"></content>';

    assert_true(d.querySelector('#shadow').offsetTop > 0, 'textarea should allow at least one insertion point');
    assert_equals(d.querySelector('#flbk').offsetTop, 0, 'Fallback content shouldn\'t be rendered');

}), 'A_09_00_01_T07');


//test progress
test(unit(function (ctx) {

	var d = newRenderedHTMLDocument(ctx);

	// create element
    var el = d.createElement('progress');
    d.body.appendChild(el);

    el.innerHTML = '' +
    	'<span id="shadow">This is a node that should be distributed</span>' +
    	'<span id="flbk">This is a fallback content</span>';

    var s = el.createShadowRoot();
    s.innerHTML = '<content select="#shadow"></content>';

    assert_true(d.querySelector('#shadow').offsetTop > 0, 'progress should allow at least one insertion point');
    assert_equals(d.querySelector('#flbk').offsetTop, 0, 'Fallback content shouldn\'t be rendered');

}), 'A_09_00_01_T08');


//test meter
test(unit(function (ctx) {

	var d = newRenderedHTMLDocument(ctx);

	// create element
    var el = d.createElement('meter');
    d.body.appendChild(el);

    el.innerHTML = '' +
    	'<span id="shadow">This is a node that should be distributed</span>' +
    	'<span id="flbk">This is a fallback content</span>';

    var s = el.createShadowRoot();
    s.innerHTML = '<content select="#shadow"></content>';

    assert_true(d.querySelector('#shadow').offsetTop > 0, 'meter should allow at least one insertion point');
    assert_equals(d.querySelector('#flbk').offsetTop, 0, 'Fallback content shouldn\'t be rendered');

}), 'A_09_00_01_T09');
</script>
</body>
</html>
